pbio/dcmotor: Fix voltage read during coast.
Technically the voltage state is not relevant during coast which
it why it was not explicitly set.

However, pbio/observer does not yet handle coast explicitly,
and just assumes that there is always a voltage. So, we have to
set the voltage to zero to make sure the observer isn't getting
a nonzero value while coasting.

Before this commit, the observer could get a slight offset while
coasting. This would cause the next command to start with an
incorrect initial condition. As a result, two identical maneuvers
used sequentially were not always equal. This is now fixed.

This also adds a test that reproduces the problem. With this fix,
it should produce a graph with two near-identical moves.