WIP: Working usb detection.
This is a lazy polling loop to debug the HAL_PCDEx_BCD_VBUSDetect
function. The "trick" seems to the that (some of) the GCCFG flags
need to be cleared, but aren't by default.

This should provide some insight to fix our (in progress) driver.

To try it, run:

from pybricks.experimental import detect
from pybricks.tools import wait

while True:

    detect()
    wait(100)